This is the last 11 commits merged into 1.
Why?
WELL, because i stupidly added .so files to one commit at the beginning and removed them later,
this is stupid BECAUSE they were 110mb each, and about 4 cpu architectures, ye, github didn't like
that, so now it hopefully fixed and i added all .so AND .a files to gitignore.